requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
fastapi==0.115.11
feedparser==6.0.11
h11==0.14.0
idna==3.10
lxml==5.3.1
lxml_html_clean==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
readability-lxml==0.8.1
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
